saparate features from netdataTable to components
